Fix #9485: ContentLength to long from int

+++ primefaces/src/main/java/org/primefaces/model/StreamedContent.java
@@ -36,7 +36,7 @@ public interface StreamedContent {
 
     String getContentEncoding();
 
-    Integer getContentLength();
+    Long getContentLength();
 
     /**
      * {@link InputStream} which will be buffered and written to the response.
